Calculates global exports by product

The function calculates global exports by product.

calc_global_exports(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  exports = IEATools::interface_industries$exports,
  Global_Exports_By_Product = "Global_Exports_By_Product"
)

Arguments

.tidy_iea_df

The .tidy_iea_df for which global exports by product need to be calculated.

flow, method, energy_type, last_stage, year, ledger_side, flow_aggregation_point, product, e_dot

See IEATools::iea_cols.

exports

The name of exports flows in the .tidy_iea_df. Default is IEATools::interface_industries$exports.

Global_Exports_By_Product

The name of the column containing global exports by product.

Value

A data frame reporting global exports by product.

Examples

tidy_AB_data %>%
calc_global_exports() %>%
print()
#> # A tibble: 4 × 8
#> # Groups:   Method, Energy.type, Last.stage, Year, Ledger.side,
#> #   Flow.aggregation.point [1]
#>   Method Energy.type Last.stage  Year Ledger.side Flow.aggrega…¹ Product Globa…²
#>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>          <chr>     <dbl>
#> 1 PCM    E           Final       2018 Supply      Total primary… Coke o…     400
#> 2 PCM    E           Final       2018 Supply      Total primary… Coking…    2000
#> 3 PCM    E           Final       2018 Supply      Total primary… Crude …    3500
#> 4 PCM    E           Final       2018 Supply      Total primary… Natura…    2500
#> # … with abbreviated variable names ¹​Flow.aggregation.point,
#> #   ²​Global_Exports_By_Product