Calculates the national production by product

The function calculates the national production by product. Included flows are flows belonging to either the R or V matrices, excluding import flows.

calc_national_production_by_product(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  V_matrix = IEATools::psut_cols$V,
  R_matrix = IEATools::psut_cols$R,
  imports = IEATools::interface_industries$imports,
  National_Production_By_Product = "National_Production_By_Product"
)

Arguments

.tidy_iea_df
The .tidy_iea_df for which national production by product needs to be calculated.

ledger_side, flow_aggregation_point, flow, e_dot
See IEATools::iea_cols.

matnames
The column name for matrices names. Default is IEATools::mat_meta_cols$matnames.

V_matrix
The name of the V matrix. Default is IEATools::psut_cols$V.

R_matrix
The name of the R matrix. Default is IEATools::psut_cols$R.

imports
The name of imports flows in the .tidy_iea_df. Default is IEATools::interface_industries$imports.

National_Production_By_Product
The name of the column containing national production for each product, by country. Default is "National_Production_By_Product".

Value

A data frame that returns the national production by product.

Details

Note: the function needs to have a column indicating matrix names added first, most likely using the IEATools::add_psut_matnames() function.

Examples

tidy_AB_data %>%
IEATools::add_psut_matnames() %>%
calc_national_production_by_product() %>%
print()

#> # A tibble: 18 × 8
#> # Groups:   Country, Method, Energy.type, Last.stage, Year, Product [18]
#>    Country Method Energy.type Last.stage  Year Product    Unit  National_Produc…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>      <chr>            <dbl>
#>  1 A       PCM    E           Final       2018 Blast fur… ktoe               850
#>  2 A       PCM    E           Final       2018 Coke oven… ktoe               400
#>  3 A       PCM    E           Final       2018 Coking co… ktoe              5000
#>  4 A       PCM    E           Final       2018 Coking co… ktoe              5000
#>  5 A       PCM    E           Final       2018 Crude oil  ktoe              8500
#>  6 A       PCM    E           Final       2018 Crude oil… ktoe              8500
#>  7 A       PCM    E           Final       2018 Electrici… ktoe              3200
#>  8 A       PCM    E           Final       2018 Heat       ktoe               250
#>  9 A       PCM    E           Final       2018 Kerosene … ktoe              1750
#> 10 A       PCM    E           Final       2018 Motor gas… ktoe              2400
#> 11 A       PCM    E           Final       2018 Natural g… ktoe              4100
#> 12 A       PCM    E           Final       2018 Natural g… ktoe              4000
#> 13 B       PCM    E           Final       2018 Blast fur… ktoe               850
#> 14 B       PCM    E           Final       2018 Coke oven… ktoe              1400
#> 15 B       PCM    E           Final       2018 Electrici… ktoe              1000
#> 16 B       PCM    E           Final       2018 Heat       ktoe               500
#> 17 B       PCM    E           Final       2018 Kerosene … ktoe              1000
#> 18 B       PCM    E           Final       2018 Motor gas… ktoe              1200