Specifies electricity and heat from nuclear

This function selects electricity and heat produced by the nuclear industry, and specifies them, so that the new product name is respectively "Electricity [from Nuclear]" and "Heat [from Nuclear]".

specify_elect_heat_nuclear(
  .tidy_iea_df,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  negzeropos = ".negzeropos"
)

Arguments

.tidy_iea_df

The .tidy_iea_df for which electricity and heat products coming from the nuclear industry need to be specified.

flow_aggregation_point, flow, e_dot, product

See IEATools::iea_cols.

transformation_processes

The name of transformation processes in the data frame. Default is IEATools::aggregation_flows$transformation_processes.

nuclear_industry

The name of the nuclear industry in the flow column of the data frame. Default is IEATools::eiou_flows$nuclear_industry.

negzeropos

Temporary column name. Default is ".netzeropos".

Value

A .tidy_iea_df with electricity and heat products specified when they come from the nuclear industry.

Examples

A_B_path <- system.file("extdata/A_B_data_full_2018_format.csv", package = "ECCTools")
IEATools::load_tidy_iea_df(A_B_path) %>%
IEATools::specify_all() %>%
specify_elect_heat_nuclear()
#> # A tibble: 102 × 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregation.poi…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>                
#>  1 A       PCM    E           Final       2018 Consumption Industry             
#>  2 A       PCM    E           Final       2018 Consumption Industry             
#>  3 A       PCM    E           Final       2018 Consumption Industry             
#>  4 A       PCM    E           Final       2018 Consumption Industry             
#>  5 A       PCM    E           Final       2018 Consumption Industry             
#>  6 A       PCM    E           Final       2018 Consumption Industry             
#>  7 A       PCM    E           Final       2018 Consumption Other                
#>  8 A       PCM    E           Final       2018 Consumption Other                
#>  9 A       PCM    E           Final       2018 Consumption Other                
#> 10 A       PCM    E           Final       2018 Consumption Other                
#> # … with 92 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>